+++ test/functional/rpc_blockchain.py
@@ -51,6 +51,8 @@ class BlockchainTest(BitcoinTestFramework):
     def set_test_params(self):
         self.setup_clean_chain = True
         self.num_nodes = 1
+        # TODO: remove -txindex. Currently required for getrawtransaction call.
+        self.extra_args = [["-txindex"]]
 
     def run_test(self):
         self.mine_chain()