+++ test/functional/wallet_abandonconflict.py
@@ -27,7 +27,8 @@ from test_framework.util import (
 class AbandonConflictTest(BitcoinTestFramework):
     def set_test_params(self):
         self.num_nodes = 2
-        self.extra_args = [["-minrelaytxfee=0.00001"], []]
+        # TODO: remove -txindex. Currently required for getrawtransaction call.
+        self.extra_args = [["-minrelaytxfee=0.00001", "-txindex"], []]
 
     def skip_test_if_missing_module(self):
         self.skip_if_no_wallet()