+++ test/functional/wallet_basic.py
@@ -23,10 +23,12 @@ from test_framework.util import (
 class WalletTest(BitcoinTestFramework):
     def set_test_params(self):
         self.num_nodes = 4
-        self.extra_args = [
-            ["-acceptnonstdtxn=1"],
-        ] * self.num_nodes
         self.setup_clean_chain = True
+        # TODO: remove -txindex. Currently required for getrawtransaction call.
+        self.extra_args = [["-acceptnonstdtxn=1"],
+                           ["-acceptnonstdtxn=1"],
+                           ["-acceptnonstdtxn=1", "-txindex"],
+                           ["-acceptnonstdtxn=1"]]
 
     def skip_test_if_missing_module(self):
         self.skip_if_no_wallet()